requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
numpy==1.21.0
scikit-learn==0.24.2
openai==0.27.0
python-dotenv==0.19.0
websockets==10.0
react==17.0.2
d3==7.0.0